short-circuit `fd_write` in the adapter
@guybedford was getting assertion errors in `State::new` due to the
debugging print statements he had added to his `cabi_realloc`
function, which called back into `fd_write`, leading to a circular
dependency where `fd_write` needed to allocate, which caused
`fd_write` to be called, etc.  The fix is easy enough: short circuit
this circularity the same way we're handling it in
`environ_sizes_get`, `clock_time_get`, etc.

If we wanted to be really paranoid, we could add this short circuit to
all the functions, but I'd be really surprised if `cabi_realloc` needs
to read from the filesystem, create directories, or other such things.
If we see that in the wild, we can revisit.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>